Rename nodeRenderers node-renderers (consistency).
